move private functions from utils back to component
