+++ client/web-sveltekit/src/lib/repo/LastCommit.svelte
@@ -16,14 +16,11 @@
 </script>
 
 <script lang="ts">
+    import { formatDistanceToNow } from 'date-fns'
+    import { capitalize } from 'lodash'
+
     import type { Avatar_User } from '$lib/Avatar.gql'
     import Avatar from '$lib/Avatar.svelte'
-    import {
-        truncateIfNeeded,
-        extractPRNumber,
-        getFirstNameAndLastInitial,
-        convertToElapsedTime,
-    } from '$lib/repo/utils'
 
     export let latestCommit: LastCommitProps
 
@@ -32,6 +29,54 @@
         avatarURL: latestCommit.author.person.avatarURL,
         displayName: latestCommit.author.person.displayName,
         username: latestCommit.author.person.name,
+    }
+
+    function getFirstNameAndLastInitial(name: string): string {
+        const names = name.split(' ')
+        if (names.length < 2) {
+            return `${capitalize(names[0].toLowerCase())}`
+        }
+        const firstName = names[0].toLowerCase()
+        const lastInitial = names[names.length - 1].charAt(0).toUpperCase()
+        return `${capitalize(firstName)} ${lastInitial}.`
+    }
+
+    function extractPRNumber(cm: string): string | null {
+        if (!hasPRNumber(cm)) {
+            return null
+        }
+        let cmWords = cm.split(' ')
+        let sha = cmWords[cmWords.length - 1]
+        return sha.slice(1, sha.length - 1)
+    }
+
+    function convertToElapsedTime(commitDateString: string): string {
+        const commitDate = new Date(commitDateString)
+        return formatDistanceToNow(commitDate, { addSuffix: true })
+    }
+
+    function truncateIfNeeded(cm: string): string {
+        cm = extractCommitMessage(cm)
+        return cm.length > 23 ? cm.substring(0, 23) + '...' : cm
+    }
+
+    function hasPRNumber(cm: string): boolean {
+        let words = cm.split(' ')
+        for (let word of words) {
+            if (/\(#(\d+)\)/.test(word)) {
+                return true
+            }
+        }
+        return false
+    }
+
+    function extractCommitMessage(cm: string): string {
+        if (!hasPRNumber(cm)) {
+            return cm
+        }
+        let splitMsg = cm.split(' ')
+        let msg = splitMsg.slice(0, splitMsg.length - 1)
+        return msg.join(' ')
     }
 
     $: commitMessageNoPR = truncateIfNeeded(latestCommit.subject)
