+++ client/web-sveltekit/src/lib/repo/utils.ts
@@ -1,5 +1,3 @@
-import { formatDistanceToNow } from 'date-fns'
-import { capitalize } from 'lodash'
 
 import { resolveRoute } from '$app/paths'
 
@@ -66,52 +64,4 @@ export async function resolveRevision(
         return revisionFromURL
     }
     return (await parent()).resolvedRevision.commitID
-}
-
-export function getFirstNameAndLastInitial(name: string): string {
-    const names = name.split(' ')
-    if (names.length < 2) {
-        return `${capitalize(names[0].toLowerCase())}`
-    }
-    const firstName = names[0].toLowerCase()
-    const lastInitial = names[names.length - 1].charAt(0).toUpperCase()
-    return `${capitalize(firstName)} ${lastInitial}.`
-}
-
-export function extractPRNumber(cm: string): string | null {
-    if (!hasPRNumber(cm)) {
-        return null
-    }
-    let cmWords = cm.split(' ')
-    let sha = cmWords[cmWords.length - 1]
-    return sha.slice(1, sha.length - 1)
-}
-
-export function convertToElapsedTime(commitDateString: string): string {
-    const commitDate = new Date(commitDateString)
-    return formatDistanceToNow(commitDate, { addSuffix: true })
-}
-
-export function truncateIfNeeded(cm: string): string {
-    cm = extractCommitMessage(cm)
-    return cm.length > 23 ? cm.substring(0, 23) + '...' : cm
-}
-
-function hasPRNumber(cm: string): boolean {
-    let words = cm.split(' ')
-    for (let word of words) {
-        if (/\(#(\d+)\)/.test(word)) {
-            return true
-        }
-    }
-    return false
-}
-
-export function extractCommitMessage(cm: string): string {
-    if (!hasPRNumber(cm)) {
-        return cm
-    }
-    let splitMsg = cm.split(' ')
-    let msg = splitMsg.slice(0, splitMsg.length - 1)
-    return msg.join(' ')
 }
